Reorganized elliptic curves module structure
