+++ math/src/elliptic_curve/mod.rs
@@ -1,6 +1,3 @@
-/// Implementation of particular cases of elliptic curves.
-pub mod curves;
-/// Implementation of EllipticCurveElement, a generic projective point in a curve.
-pub mod element;
-/// Common behaviour for Elliptic curves.
-pub mod traits;
+pub mod edwards;
+pub mod montgomery;
+pub mod short_weierstrass;
