+++ math/src/elliptic_curve/short_weierstrass/mod.rs
@@ -1,0 +1,6 @@
+/// Implementation of particular cases of elliptic curves.
+pub mod curves;
+/// Implementation of EllipticCurveElement, a generic projective point in a curve.
+pub mod element;
+/// Common behaviour for Elliptic curves.
+pub mod traits;
